BUG-121788 HBASE-22012 Prevent DISABLE quota violation policy from disabling violation policy
Space quotas has a feature which intends to avoid enacting a space quota
violation policy when only a subset of the Regions for that Table have
reported their space usage (under the assumption that we cannot make an
informed decision if we do not include all regions in our calculations).

This had the unintended side-effect, when a table is disabled as a part
of a violation policy, of causing the regions for that table to not be
reported which disables the violation policy and enables the table.

Need to make sure that when a table is disabled because of a violation
policy that the code does not automatically move that table out of
violation because region sizes are not being reported (because those
regions are not open).

elserj: Had to replace some usages of Optional that don't exist in branch-2.1

Closes #572

Signed-off-by: Josh Elser <elserj@apache.org>
(cherry picked from commit 23adf23a4fb896d9f9a6f1903417cf89e4a9bf5b)
Change-Id: I6dac07a90aae75bf0ab95d8e953919fef3713da4